chore: rearrange header file
