+++ sensing/pointcloud_preprocessor/include/pointcloud_preprocessor/blockage_diag/blockage_diag_nodelet.hpp
@@ -1,4 +1,4 @@
-// Copyright 2021 Tier IV, Inc.
+// Copyright 2022 Tier IV, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -24,12 +24,10 @@
 
 #include <diagnostic_msgs/msg/diagnostic_array.hpp>
 #include <sensor_msgs/msg/point_cloud2.hpp>
-#include <tier4_debug_msgs/msg/float32_multi_array_stamped.hpp>
+#include <std_msgs/msg/header.hpp>
 #include <tier4_debug_msgs/msg/float32_stamped.hpp>
 
 #include <cv_bridge/cv_bridge.h>
-#include <pcl/filters/voxel_grid.h>
-#include <pcl/search/pcl_search.h>
 
 #include <string>
 #include <vector>
@@ -38,16 +36,6 @@ namespace pointcloud_preprocessor
 {
 using diagnostic_updater::DiagnosticStatusWrapper;
 using diagnostic_updater::Updater;
-enum ReturnType : uint8_t {
-  INVALID = 0,
-  SINGLE_STRONGEST,
-  SINGLE_LAST,
-  DUAL_STRONGEST_FIRST,
-  DUAL_STRONGEST_LAST,
-  DUAL_WEAK_FIRST,
-  DUAL_WEAK_LAST,
-  DUAL_ONLY,
-};
 
 class BlockageDiagComponent : public pointcloud_preprocessor::Filter
 {
@@ -67,15 +55,12 @@ protected:
 
 private:
   void onBlockageChecker(DiagnosticStatusWrapper & stat);
-  void onSkyBlockageChecker(DiagnosticStatusWrapper & stat);
   Updater updater_{this};
   uint vertical_bins_;
   std::vector<double> angle_range_deg_;
-  std::vector<double> distance_range_;
   uint horizontal_ring_id_ = 12;
-  float ground_blockage_threshold_ = 0.1;
+  float blockage_ratio_threshold_;
   float ground_blockage_ratio_ = -1.0f;
-  float sky_blockage_threshold_ = 0.2f;
   float sky_blockage_ratio_ = -1.0f;
   std::vector<float> ground_blockage_range_deg_ = {0.0f, 0.0f};
   std::vector<float> sky_blockage_range_deg_ = {0.0f, 0.0f};
@@ -83,7 +68,7 @@ private:
   uint ground_blockage_count_ = 0;
   uint sky_blockage_count_ = 0;
   uint blockage_count_threshold_;
-  std::string lidar_model_ = "Pandar40P";
+  std::string lidar_model_;
 
 public:
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
@@ -91,27 +76,5 @@ public:
 };
 
 }  // namespace pointcloud_preprocessor
-
-namespace return_type_cloud
-{
-struct PointXYZIRADT
-{
-  PCL_ADD_POINT4D;
-  float intensity;
-  std::uint16_t ring;
-  float azimuth;
-  float distance;
-  std::uint8_t return_type;
-  double time_stamp;
-  EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-} EIGEN_ALIGN16;
-
-}  // namespace return_type_cloud
-
-POINT_CLOUD_REGISTER_POINT_STRUCT(
-  return_type_cloud::PointXYZIRADT,
-  (float, x, x)(float, y, y)(float, z, z)(float, intensity, intensity)(std::uint16_t, ring, ring)(
-    float, azimuth, azimuth)(float, distance, distance)(double, time_stamp, time_stamp)(
-    std::uint8_t, return_type, return_type))
 
 #endif  // POINTCLOUD_PREPROCESSOR__BLOCKAGE_DIAG__BLOCKAGE_DIAG_NODELET_HPP_
