+++ sensing/pointcloud_preprocessor/include/pointcloud_preprocessor/outlier_filter/dual_return_outlier_filter_nodelet.hpp
@@ -39,17 +39,6 @@ namespace pointcloud_preprocessor
 using diagnostic_updater::DiagnosticStatusWrapper;
 using diagnostic_updater::Updater;
 
-enum ReturnType : uint8_t {
-  INVALID = 0,
-  SINGLE_STRONGEST,
-  SINGLE_LAST,
-  DUAL_STRONGEST_FIRST,
-  DUAL_STRONGEST_LAST,
-  DUAL_WEAK_FIRST,
-  DUAL_WEAK_LAST,
-  DUAL_ONLY,
-};
-
 std::unordered_map<std::string, uint8_t> roi_mode_map_ = {
   {"No_ROI", 0},
   {"Fixed_xyz_ROI", 1},
@@ -100,27 +89,5 @@ public:
 };
 
 }  // namespace pointcloud_preprocessor
-
-namespace return_type_cloud
-{
-struct PointXYZIRADT
-{
-  PCL_ADD_POINT4D;
-  float intensity;
-  std::uint16_t ring;
-  float azimuth;
-  float distance;
-  std::uint8_t return_type;
-  double time_stamp;
-  EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-} EIGEN_ALIGN16;
-
-}  // namespace return_type_cloud
-
-POINT_CLOUD_REGISTER_POINT_STRUCT(
-  return_type_cloud::PointXYZIRADT,
-  (float, x, x)(float, y, y)(float, z, z)(float, intensity, intensity)(std::uint16_t, ring, ring)(
-    float, azimuth, azimuth)(float, distance, distance)(std::uint8_t, return_type, return_type)(
-    double, time_stamp, time_stamp))
 
 #endif  // POINTCLOUD_PREPROCESSOR__OUTLIER_FILTER__DUAL_RETURN_OUTLIER_FILTER_NODELET_HPP_
