+++ sensing/pointcloud_preprocessor/launch/blockage_diag.launch.xml
@@ -1,9 +1,6 @@
 <launch>
-  <arg name="input_topic_name" default="rectified/pointcloud_ex" />
+  <arg name="input_topic_name" default="pointcloud_raw_ex" />
   <arg name="output_topic_name" default="blockage_diag/pointcloud" />
-
-  <arg name="input_frame" default="velodyne" />
-  <arg name="output_frame" default="velodyne" />
 
   <node pkg="pointcloud_preprocessor" exec="blockage_diag_node" name="blockage_diag">
     <remap from="input" to="$(var input_topic_name)" />
